+++ client/app/account/user/advanced/advanced.routing.js
@@ -10,6 +10,7 @@ angular
         url: '/advanced',
         templateUrl: 'account/user/advanced/user-advanced.html',
         controller: 'UserAccount.controllers.advanced',
+        controllerAs: 'advancedCtrl',
         translations: ['../'],
       });
     }